mcp-server/apigee-mcp-server/archive/requirements.txt
fastapi
fastmcp
google-auth>=2.26.1,<3.0dev
google-auth-httplib2>=0.2.0,<1.0.0
PyYAML>=6.0.2
requests>=2.32.4
uvicorn>=0.34.0